Install OTP on Windows to a local directory
It is especially useful not to install things globally on self-hosted
runner as these may not be running in isolation.

By installing OTP into a known location we could remove the code that
was checking where exactly it was installed.

This PR also removes the `echo $INSTALL_DIR_FOR_OTP >> $GITHUB_ENV` (and
the PowerShell equivalent) as it never worked. The variable name should
have been `$GITHUB_PATH`, not `$GITHUB_ENV`. Again, no longer necessary
as we install OTP to a known location.