Link to the elixir/erlang compatibility page
Currently when a user encounters an incompatibility between the
requested elixir and erlang versions it isn't clear how to figure out
what the compatible versions are. This is the current error as shown to users:

> Error: Requested Elixir / Erlang/OTP version (1.11.3 / 24.2) not found in version list (did you check Compatibility between Elixir and Erlang/OTP?)

Example failure: https://github.com/ScenicFramework/scenic/pull/318/checks

This PR adds a link to
https://hexdocs.pm/elixir/compatibility-and-deprecations.html so user's
can easily see what versions are compatible.